iomap: Introduce IOMAP_ENCODED
An encoded extent must be read completely. Make the bio just as a
regular bio and let filesystem deal with the rest of the extent.
However, the filesystem must be informed that the bio to be read is
encoded and the offset from which the encoded extent starts. So, pass
the iomap associated with the bio while calling submit_io. Save the
previous iomap (associated with the bio being submitted) in prev in
order to submit when the iomap changes.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>